Outputting data as a table and working on transofmring approach to using signals
