+++ src/app/app.component.ts
@@ -11,38 +11,4 @@ import { InvestmentResultsComponent } from "./investment-results/investment-resu
   imports: [HeaderComponent, UserInputComponent, InvestmentResultsComponent],
   templateUrl: './app.component.html',
 })
-export class AppComponent {
-
-  resultsData = signal<{
-    year: number;
-    interest: number;
-    valueEndOfYear: number;
-    annualInvestment: number;
-    totalInterest: number;
-    totalAmountInvested: number;
-  }[] | undefined>(undefined);
-  onCalculateInvestmentResults(data: InvestmentInput) {
-    const { initialInvestment, annualInvestment, expectedReturn, duration } = data;
-    const annualData = [];
-    let investmentValue = initialInvestment;
-  
-    for (let i = 0; i < duration; i++) {
-      const year = i + 1;
-      const interestEarnedInYear = investmentValue * (expectedReturn / 100);
-      investmentValue += interestEarnedInYear + annualInvestment;
-      const totalInterest =
-        investmentValue - annualInvestment * year - initialInvestment;
-      annualData.push({
-        year: year,
-        interest: interestEarnedInYear,
-        valueEndOfYear: investmentValue,
-        annualInvestment: annualInvestment,
-        totalInterest: totalInterest,
-        totalAmountInvested: initialInvestment + annualInvestment * year,
-      });
-    }
-  
-    //return annualData;
-    this.resultsData.set(annualData);
-}
-}
+export class AppComponent { }
