+++ src/app/investment-results/investment-results.component.html
@@ -1,4 +1,4 @@
-@if (!results()) {
+@if (!results) {
     <p>Please enter some values and press "Calculate" to see the results.</p>
 } @else {
     <table>
@@ -12,7 +12,7 @@
             </tr>
         </thead>
         <tbody>
-            @for (result of results(); track result.year) {
+            @for (result of results; track result.year) {
                 <tr>
                     <td>{{ result.year }}</td>
                     <td>{{ result.valueEndOfYear | currency }}</td>
