+++ src/app/investment-results/investment-results.component.ts
@@ -1,5 +1,6 @@
-import { Component, input, } from '@angular/core';
+import { Component,inject } from '@angular/core';
 import { CurrencyPipe } from '@angular/common';
+import { InvestmentService } from '../investment.service';
 
 @Component({
   selector: 'app-investment-results',
@@ -9,12 +10,10 @@ import { CurrencyPipe } from '@angular/common';
   styleUrl: './investment-results.component.css'
 })
 export class InvestmentResultsComponent {
-  results = input< {
-    year: number;
-    interest: number;
-    valueEndOfYear: number;
-    annualInvestment: number;
-    totalInterest: number;
-    totalAmountInvested: number;
-  }[]>();
+  private investmentService: inject(InvestmentService);
+
+  get results(){
+    return this.investmentService.resultsData;
+  }
 }
+
